requirements.txt
Django==4.1.7
django-environ==0.10.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
drf-yasg==1.21.5
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
python-dotenv==1.0.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
requests==2.29.0
pydantic==1.10.9
marshmallow==3.19.0
marshmallow_dataclass==8.5.14
pytest~=7.4.0